[FEATURE] Changes in the commander. Use only file supplied as an argument

+++ commander/main.cpp
@@ -8,73 +8,66 @@
 #include "PECommander.h"
 
 #define TITLE "BearCommander"
+#define MINBUF 0x200
 
 using namespace std;
 
-QString getFileName()
+FileView* tryLoading(QString &fName)
 {
-    QTextStream qtin(stdin, QIODevice::ReadOnly);
-    QString fName;
-
-    int trials = 3;
+	FileView *fileView = NULL;
+    bufsize_t maxMapSize = FILE_MAXSIZE;
     do {
-        printf("Path to executable: ");
-        fName = qtin.readLine();
-        if (QFile::exists(fName)) break;
-
-        std::cerr << "No such file! Remaining attempts: " << std::dec << trials << std::endl;
-        trials--;
-
-    } while (trials);
-
-    return fName;
+		try {
+			fileView = new FileView(fName, maxMapSize);
+		} catch (BufferException &e1) {
+			std::cerr << "[ERROR] " << e1.what() << std::endl;
+			maxMapSize = static_cast<bufsize_t>(cmd_util::readNumber("Try again with size (hex): ", true));
+			if (maxMapSize == 0) break;
+		}
+	} while (!fileView);
+	
+	return fileView;
 }
 
 int main(int argc, char *argv[])
 {
-    std::cout << "Bearparser version: " <<  BEARPARSER_VERSION << std::endl;
     QCoreApplication app(argc, argv);
 
     ExeFactory::init();
     ExeCmdContext exeContext;
     PECommander commander(&exeContext);
 
-    QString fName;
-    if (argc < 2) {
-        fName = getFileName();
-    } else {
-        fName = QString(argv[1]);
-    }
+	if (argc < 2) {
+		std::cout << "Bearparser version: " <<  BEARPARSER_VERSION << "\n";
+		std::cout << "Args: <PE file>\n";
+		commander.printHelp();
+		return 0;
+	}
+	
+	
+	int status = 0;
+	QString fName = QString(argv[1]);
+	
     try {
-        FileView *fileView = NULL;
-        bufsize_t maxMapSize = FILE_MAXSIZE;
-        do {
-            try {
-                fileView = new FileView(fName, maxMapSize);
-            
-            } catch (BufferException &e1) {
-                std::cerr << "[ERROR] " << e1.what() << std::endl;
-                maxMapSize = static_cast<bufsize_t>(cmd_util::readNumber("Try again with size (hex): ", true));
-            }
-        } while (fileView == NULL);
+        FileView* fileView = tryLoading(fName);
+		if (!fileView) return -1;
 
         ExeFactory::exe_type exeType = ExeFactory::findMatching(fileView);
         if (exeType == ExeFactory::NONE) {
-           fprintf(stderr, "Type not supported\n");
+           std::cerr << "Type not supported\n";
            ExeFactory::destroy();
            return 1;
         }
         
         std::cout << "Type: " << ExeFactory::getTypeName(exeType).toStdString() << std::endl;
-        const bufsize_t MINBUF = 0x200;
         bufsize_t readableSize = fileView->getContentSize();
         bufsize_t allocSize = (readableSize < MINBUF) ? MINBUF : readableSize;
 
-        std::cerr << "Buffering..." << std::endl;
+        std::cout << "Buffering..." << std::endl;
         ByteBuffer *buf = new ByteBuffer(fileView, 0, allocSize);
-        delete fileView;
+        delete fileView; fileView = NULL; //the view is no longer needed
 
-        std::cerr << "Parsing executable..." << std::endl;
+        std::cout << "Parsing executable..." << std::endl;
         Executable *exe = ExeFactory::build(buf, exeType);
 
         exeContext.setExe(exe);
@@ -82,12 +75,14 @@ int main(int argc, char *argv[])
 
         delete exe;
         delete buf;
-
+		
+		std::cout << "Bye!" << std::endl;
+		
     } catch (CustomException &e) {
         std::cerr << "[ERROR] " << e.what() << std::endl;
+		status = -1;
     }
-    std::cout << "Bye!" << std::endl;
     ExeFactory::destroy();
-    return 0;
+    return status;
 }
